Extract regional functions into helper module
